Que sonrisa maléfica tiene Gabriel cuando pregunta fronteras de conjuntos
