+++ src/pages/AfterLogin.js
@@ -1,3 +1,4 @@
+/*
 import React from 'react';
 import flecha from '../assets/flecha.png';
 import { useNavigate } from 'react-router-dom';
@@ -26,10 +27,11 @@ function AfterLogin() {
         />
         </div>
       </div>
-      {/* Imagen del futbolista */}
+      {/* Imagen del futbolista }
       <img src={equipo} alt="Equipo" className="equipo-image" />
     </div>
   );
 }
 
 export default AfterLogin;
+*/
