+++ src/styles/pages/home.css
@@ -1,85 +1,133 @@
-.home {
-  color: aliceblue;
-  display: flex;
-  flex-direction: column; /* Por defecto, columna */
-  height: 100vh; /* Toma toda la altura de la ventana del navegador */
-  align-items: center; /* Centra verticalmente los elementos */
-  justify-content: center;
+
+.information-body {
+    background-color: #173b4c; 
+    font-family: 'Roboto Serif', serif; 
+    font-weight: bold; 
+    text-align: center; 
+    margin: 0; 
+    padding: 20px; 
+    display: flex; 
+    justify-content: center; 
+    align-items: flex-start; 
+    flex-direction: column; 
+    min-height: 100vh; 
+    box-sizing: border-box; 
 }
 
-@media (min-width: 768px) {
-  .home {
-      flex-direction: row; /* En pantallas anchas, cambia a fila */
-  }
+.information-main-container {
+    display: flex;
+    justify-content: space-between; 
+    align-items: center; 
+    width: 100vw; 
+    max-width: 1200px; 
+    margin-bottom: 3%; 
+    flex-wrap: wrap; 
+    margin-top: 25%; 
 }
 
-.line {
-  width: 100%;
-  height: 2px; /* Línea horizontal por defecto */
-  background-color: #ccc;
-  margin: 20px 0;
+.information-image {
+    max-width: 100%; 
+    height: auto; 
+    border-radius: 8px; 
 }
 
-@media (min-width: 768px) {
-  .line {
-      width: 2px; /* Línea vertical en pantallas anchas */
-      height: 50%;
-      margin: 0 20px;
-  }
+.information-steps-container {
+    width: 80vw; 
+    max-width: 1200px; 
+    background-color: #173b4c; 
+    border-radius: 50px; 
+    padding: 40px 20px; 
+    margin-bottom: 20px; 
+    margin-top: 10px; 
 }
 
-.welcome {
-  padding: 20px;
-  max-width: 800px;
+.information-steps {
+    display: flex; 
+    justify-content: space-between; 
+    flex-wrap: wrap; 
 }
 
-.actions {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+.information-step {
+    flex: 1; 
+    margin: 10px;
+    background-color: #ffffff; 
+    padding: 20px; 
+    border-radius: 8px; 
+    text-align: center;
+    display: flex;
+    flex-direction: column; 
+    align-items: center; 
 }
 
-.actions button {
-  margin-top: 10px;
+.information-step-title {
+    font-size: 24px; 
+    font-weight: bold; 
+    color: #1DB4BB; 
+    margin-bottom: 10px; 
+    text-align: center; 
 }
 
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5); /* Fondo semitransparente */
-  z-index: 999; /* Asegura que el overlay esté sobre todo */
+.information-step-image {
+    max-width: 50px; 
+    height: auto;
+    margin-bottom: 10px; 
 }
 
-.form-popup {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: white;
-  padding: 20px;
-  border-radius: 5px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+.information-step p,
+.information-step ul {
+    text-align: center;
 }
 
-.home h1 {
-  font-family: 'Inter', sans-serif;
-  font-weight: 400; /* Misma fuente en negrita */
-  font-size: 50px; /* Mismo tamaño de fuente */
-  letter-spacing: 2px; /* Mismo espaciado entre letras */
-  color: #1DB4BB; /* Mismo color de texto */
-  text-align: center; /* Centra horizontalmente el texto */
-  margin-bottom: 0;
+.information-h1 {
+    font-family: 'Inter', sans-serif;
+    font-weight: 700;
+    font-size: 48px; 
+    letter-spacing: 1px; 
+    color: #6B9080; 
+    text-align: center; 
+    line-height: 1.2; 
 }
 
-.home p {
-  font-family: 'Inter', sans-serif;
-  font-weight: 200; /* Misma fuente en negrita */
-  font-size: 25px; /* Mismo tamaño de fuente */
-  letter-spacing: 2px; /* Mismo espaciado entre letras */
-  color: #1DB4AA; /* Mismo color de texto */
-  text-align: center; /* Centra horizontalmente el texto */
-  margin-bottom: 0;
+.information-h2 {
+    font-family: 'Inter', sans-serif;
+    font-weight: 400; 
+    font-size: 40px; 
+    letter-spacing: 1px;
+    color: #6B9080; 
+    text-align: center; 
+    margin-bottom: 10px;
+    margin-top: 10px;
+}
+
+.information-final-text {
+    font-size: 20px; 
+    color: #ffffff;
+    margin-bottom: 20px; 
+}
+
+.button-container {
+    display: flex;
+    justify-content: center; 
+    width: 100%; 
+}
+
+.information-button {
+    cursor: pointer;
+    width: 100%;
+    max-width: 335px; 
+    height: 48px; 
+    border: 0; 
+    border-radius: 100px; 
+    background-color: #1DB4BB; 
+    color: #ffffff;
+    font-size: 18px;
+    font-family: "Inter"; 
+    font-weight: 700; 
+    outline: none; 
+    margin-top: 20px; 
+    transition: background-color 0.3s ease; 
+}
+
+.information-button:hover {
+    background-color: #14918a; 
 }
