+++ src/styles/pages/organizations.css
@@ -1,0 +1,85 @@
+.home {
+  color: aliceblue;
+  display: flex;
+  flex-direction: column; /* Por defecto, columna */
+  height: 100vh; /* Toma toda la altura de la ventana del navegador */
+  align-items: center; /* Centra verticalmente los elementos */
+  justify-content: center;
+}
+
+@media (min-width: 768px) {
+  .home {
+      flex-direction: row; /* En pantallas anchas, cambia a fila */
+  }
+}
+
+.line {
+  width: 100%;
+  height: 2px; /* Línea horizontal por defecto */
+  background-color: #ccc;
+  margin: 20px 0;
+}
+
+@media (min-width: 768px) {
+  .line {
+      width: 2px; /* Línea vertical en pantallas anchas */
+      height: 50%;
+      margin: 0 20px;
+  }
+}
+
+.welcome {
+  padding: 20px;
+  max-width: 800px;
+}
+
+.actions {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.actions button {
+  margin-top: 10px;
+}
+
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5); /* Fondo semitransparente */
+  z-index: 999; /* Asegura que el overlay esté sobre todo */
+}
+
+.form-popup {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: white;
+  padding: 20px;
+  border-radius: 5px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+}
+
+.home h1 {
+  font-family: 'Inter', sans-serif;
+  font-weight: 400; /* Misma fuente en negrita */
+  font-size: 50px; /* Mismo tamaño de fuente */
+  letter-spacing: 2px; /* Mismo espaciado entre letras */
+  color: #1DB4BB; /* Mismo color de texto */
+  text-align: center; /* Centra horizontalmente el texto */
+  margin-bottom: 0;
+}
+
+.home p {
+  font-family: 'Inter', sans-serif;
+  font-weight: 200; /* Misma fuente en negrita */
+  font-size: 25px; /* Mismo tamaño de fuente */
+  letter-spacing: 2px; /* Mismo espaciado entre letras */
+  color: #1DB4AA; /* Mismo color de texto */
+  text-align: center; /* Centra horizontalmente el texto */
+  margin-bottom: 0;
+}
